Quick fix to donation demo and fixed package scripts.
