+++ examples/donation-demo/src/components/SubBox/SubBox.module.css
@@ -23,28 +23,6 @@
   font-weight: 700;
 }
 
-.Wrapper button {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 8px;
-  background: #eb5428;
-  color: #fff;
-  height: 52px;
-  width: 100%;
-  justify-content: center;
-  border-radius: 10px;
-  font-size: 16px;
-  font-weight: 500;
-  font-family: var(--font-sf);
-  padding: 0 32px;
-  border: none;
-  box-shadow: none;
-  outline: none;
-  margin-top: 20px;
-  cursor: pointer;
-}
-
 @media screen and (max-width: 480px) {
   .Wrapper {
     width: 100%;
