+++ examples/donation-demo/src/components/SubscribeButton/SubscribeButton.module.css
@@ -1,0 +1,21 @@
+.SubscribeButton {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 8px;
+  background: #eb5428;
+  color: #fff;
+  height: 52px;
+  width: 100%;
+  justify-content: center;
+  border-radius: 10px;
+  font-size: 16px;
+  font-weight: 500;
+  font-family: var(--font-sf);
+  padding: 0 32px;
+  border: none;
+  box-shadow: none;
+  outline: none;
+  margin-top: 20px;
+  cursor: pointer;
+}
