+++ examples/donation-demo/src/styles/Home.module.css
@@ -12,6 +12,7 @@
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   gap: 24px;
+  align-items: flex-start;
 }
 
 .HeroWrapper {
