+++ package.json
@@ -8,9 +8,9 @@
     "changeset:release": "pnpm build && changeset publish",
     "changeset:version": "changeset version && pnpm install --lockfile-only",
     "clean": "npm-run-all -p clean:*",
-    "clean:next": "rimraf --g \"./**/.next\"",
-    "clean:node-modules": "rimraf --g \"**/node_modules\"",
-    "clean:dist": "rimraf --g \"./**/dist\"",
+    "clean:next": "rimraf -g \"./**/.next\"",
+    "clean:node-modules": "rimraf -g \"**/node_modules\"",
+    "clean:dist": "rimraf -g \"./**/dist\"",
     "dev": "turbo run dev",
     "dev:main": "turbo run dev --filter=!./examples/*",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
